Update autosetup.sh
unfucked autosetup script. Make sure not to just cat into bash, actually give this a 777 or smth, and then execute it as root for best effect. (ik, skill issue on my part.) If unable to run specific nics, or having no dice with the rtl drivers, google for documentation on correcting that issue. It was a previous issue with the kismet rolling release that I think got fixed like a month ago. Not really sure. Sometimes it works sometimes not. Computers are weird.